% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct1.0_target_tt.R
\name{fct1.0_target_tt}
\alias{fct1.0_target_tt}
\title{Function that identifies patients who received the treatments we want to study}
\usage{
fct1.0_target_tt(
  dv,
  line = 1,
  tt_xp1,
  tt_xp2 = NULL,
  tt_xp3 = NULL,
  tt_xp4 = NULL,
  tt_xp5 = NULL,
  tt_ct1,
  tt_ct2 = NULL,
  tt_ct3 = NULL,
  tt_ct4 = NULL,
  tt_ct5 = NULL,
  design = "A"
)
}
\arguments{
\item{dv}{to be completed}

\item{line}{to be completed}

\item{tt_xp1}{to be completed}

\item{tt_xp2}{to be completed}

\item{tt_xp3}{to be completed}

\item{tt_xp4}{to be completed}

\item{tt_xp5}{to be completed}

\item{tt_ct1}{to be completed}

\item{tt_ct2}{to be completed}

\item{tt_ct3}{to be completed}

\item{tt_ct4}{to be completed}

\item{tt_ct5}{to be completed}

\item{design}{to be completed}
}
\value{
to be completed
}
\description{
Function that identifies patients who received the treatments we want to study
}
\examples{
#to be completed


}
