% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2.0_identify_prior_tt.R
\name{fct2.0_identify_prior_tt}
\alias{fct2.0_identify_prior_tt}
\title{Identify patients who received prior drugs before T0 (for instance, prior HT when it is not an exclusion criterion)}
\usage{
fct2.0_identify_prior_tt(dv, dd, dt)
}
\arguments{
\item{dv}{to be completed}

\item{dd}{to be completed}

\item{dt}{to be completed}
}
\value{
to be completed
}
\description{
Identify patients who received prior drugs before T0 (for instance, prior HT when it is not an exclusion criterion)
}
\examples{
#to be completed


}
