% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2.1_identify_prior_tt_start.R
\name{fct2.1_identify_prior_tt_start}
\alias{fct2.1_identify_prior_tt_start}
\title{Function that identifies patients with treatment/surgery/histology/procedure within the chosen delay, before T0}
\usage{
fct2.1_identify_prior_tt_start(dv, dd, dt, delay)
}
\arguments{
\item{dv}{to be completed}

\item{dd}{to be completed}

\item{dt}{to be completed}

\item{delay}{to be completed}
}
\value{
to be completed
}
\description{
Function that identifies patients with treatment/surgery/histology/procedure within the chosen delay, before T0
}
\examples{
#to be completed

}
