% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2.2_identify_prior_tt_start_end.R
\name{fct2.2_identify_prior_tt_start_end}
\alias{fct2.2_identify_prior_tt_start_end}
\title{Fonction qui filtre sur les critères relatifs aux TRT avec date de début et date de fin}
\usage{
fct2.2_identify_prior_tt_start_end(dv, dd, stdt, endt, delay)
}
\arguments{
\item{dv}{to be completed}

\item{dd}{to be completed}

\item{stdt}{to be completed}

\item{endt}{to be completed}

\item{delay}{to be completed}
}
\value{
to be completed
}
\description{
Fonction qui filtre sur les critères relatifs aux TRT avec date de début et date de fin
}
\examples{
#to be completed


}
