% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2.3_identify_tt_grace.R
\name{fct2.3_identify_tt_grace}
\alias{fct2.3_identify_tt_grace}
\title{Fonction qui filtre sur la grace period}
\usage{
fct2.3_identify_tt_grace(dv, dd, diag_dt, delay1, delay2)
}
\arguments{
\item{dv}{to be completed}

\item{dd}{to be completed}

\item{diag_dt}{to be completed}

\item{delay1}{to be completed}

\item{delay2}{to be completed}
}
\value{
to be completed
}
\description{
Fonction qui filtre sur la grace period
}
\examples{
#to be completed

}
