% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2.3_identify_tt_window.R
\name{fct2.3_identify_tt_window}
\alias{fct2.3_identify_tt_window}
\title{Function that filters according to time window eligibility}
\usage{
fct2.3_identify_tt_window(dv, dd, diag_dt, delay1, delay2)
}
\arguments{
\item{dv}{to be completed}

\item{dd}{to be completed}

\item{diag_dt}{to be completed}

\item{delay1}{to be completed}

\item{delay2}{to be completed}
}
\value{
to be completed
}
\description{
Function that filters according to time window eligibility
}
\examples{
#to be completed

}
