% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct3.0_target_eligibility.R
\name{fct3.0_target_eligibility}
\alias{fct3.0_target_eligibility}
\title{Function that filters patients based on baseline inclusion and/or exclusion criteria}
\usage{
fct3.0_target_eligibility(dd, criter_in = NULL, criter_ex = NULL, ARM = "ARM")
}
\arguments{
\item{dd}{to be completed}

\item{criter_in}{to be completed}

\item{criter_ex}{to be completed}

\item{ARM}{to be completed}
}
\value{
to be completed
}
\description{
Function that filters patients based on baseline inclusion and/or exclusion criteria
}
\examples{
#to be completed

}
