% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct4.0_propensity_weights.R
\name{fct4.0_propensity_weights}
\alias{fct4.0_propensity_weights}
\title{Return propensity scores and weights}
\usage{
fct4.0_propensity_weights(dd, confounders, arm, w_method)
}
\arguments{
\item{dd}{a data frame}

\item{confounders}{a list of confounding factors}

\item{arm}{the 0/1 binary experimental arm}

\item{w_method}{can take the value "iptwate" (IPTW-ATE), "sw" (SIPTW), or "psow" (PSOW)}
}
\value{
to be completed
}
\description{
Return propensity scores and weights
}
\examples{
#to be completed

}
