% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct4.1_propensity_overlap.R
\name{fct4.1_propensity_overlap}
\alias{fct4.1_propensity_overlap}
\title{Descriptive analysis of the propensity scores and weights, and assessment of the common support}
\usage{
fct4.1_propensity_overlap(dd, arm, ps, weights)
}
\arguments{
\item{dd}{a data frame}

\item{arm}{the 0/1 binary variable depending on whether the patients received either experimental trt or control trt}

\item{ps}{a variable that contains propensity score}

\item{weights}{a variable that contains a weight}
}
\value{
Summary statistics of propensity scores and weights, and return the distribution of propensity scores density plot
}
\description{
Descriptive analysis of the propensity scores and weights, and assessment of the common support
}
\examples{
#to be completed

}
