% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct4.3_classic_univ_table.R
\name{fct4.3_classic_univ_table}
\alias{fct4.3_classic_univ_table}
\title{to be completed}
\usage{
fct4.3_classic_univ_table(
  dd,
  n_midd = NULL,
  confounders,
  arm,
  vname_weights = NULL,
  label = NULL
)
}
\arguments{
\item{dd}{to be completed}

\item{n_midd}{to be completed}

\item{confounders}{to be completed}

\item{arm}{to be completed}

\item{vname_weights}{to be completed}

\item{label}{to be completed}
}
\value{
to be completed
}
\description{
to be completed
}
\examples{
#to be completed

}
