% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct5.1_univar_model.R
\name{fct5.1_univar_model}
\alias{fct5.1_univar_model}
\title{Univariable/Naïve analysis}
\usage{
fct5.1_univar_model(dd, fu_time, fu_event, arm)
}
\arguments{
\item{dd}{a data frame}

\item{fu_time}{the follow-up time variable for right censored data}

\item{fu_event}{the status event indicator variable: 0 = alive/1 = dead, or T=death/F=alive or 1=alive/2=death}

\item{arm}{the 0/1 binary experimental arm variable}
}
\value{
the estimate of the crude treatment effect from univariable Cox model
}
\description{
Univariable/Naïve analysis
}
\examples{
#to be completed

}
