% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct5.2_weighted_model.R
\name{fct5.2_weighted_model}
\alias{fct5.2_weighted_model}
\title{Weighted analysis}
\usage{
fct5.2_weighted_model(dd, fu_time, fu_event, arm, vname_weights)
}
\arguments{
\item{dd}{a data frame}

\item{fu_time}{the follow-up time variable for right censored data}

\item{fu_event}{the status event indicator variable: 0 = alive/1 = dead, or T=death/F=alive or 1=alive/2=death}

\item{arm}{the 0/1 binary treatment arm variable}

\item{vname_weights}{the weights variable: could be "iptwate", "sw" or "psow" if propensity_weights function was used, or other}
}
\value{
the estimates of the marginal treatment effect from weighted Cox model
}
\description{
Weighted analysis
}
\examples{
#to be completed

}
