% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct5.3_multivar_model.R
\name{fct5.3_multivar_model}
\alias{fct5.3_multivar_model}
\title{Multivariable analysis}
\usage{
fct5.3_multivar_model(dd, confounders, fu_time, fu_event, arm)
}
\arguments{
\item{dd}{a data frame}

\item{confounders}{a list of confounding factors}

\item{fu_time}{the follow-up time variable for right censored data}

\item{fu_event}{the status event indicator variable: 0 = alive/1 = dead, or T=death/F=alive or 1=alive/2=death}

\item{arm}{the 0/1 binary experimental arm variable}
}
\value{
the estimates of the conditional treatment effect from multivariable Cox model
}
\description{
Multivariable analysis
}
\examples{
#to be completed

}
