% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct6.0_Pool.R
\name{fct6.0_Pool}
\alias{fct6.0_Pool}
\title{function that pool HRs estimates according to Bernard and Rubin's rules
cf : https://bookdown.org/mwheymans/bookmi/rubins-rules.html}
\usage{
fct6.0_Pool(x, n_events, n_parameters, alpha = 0.05)
}
\arguments{
\item{x}{to be completed}

\item{n_events}{to be completed}

\item{n_parameters}{to be completed}

\item{alpha}{to be completed}
}
\value{
to be completed
}
\description{
function that pool HRs estimates according to Bernard and Rubin's rules
cf : https://bookdown.org/mwheymans/bookmi/rubins-rules.html
}
\examples{
data(cars)
m1 <- lm(speed ~ dist, data = cars[sample(30, replace = TRUE),])
m2 <- lm(speed ~ dist, data = cars[sample(30, replace = TRUE),])
m3 <- lm(speed ~ dist, data = cars[sample(30, replace = TRUE),])
m4 <- lm(speed ~ dist, data = cars[sample(30, replace = TRUE),])
m5 <- lm(speed ~ dist, data = cars[sample(30, replace = TRUE),])
m <- list(m1, m2, m3, m4, m5)

msum <- lapply(m, broom::tidy) \%>\%
  bind_rows() \%>\%
  filter(term == "dist")
fct6.0_Pool(msum, n_events = 30, n_parameters = 2)

}
