% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetesme-package.R
\docType{package}
\name{targetesme-package}
\alias{targetesme}
\alias{targetesme-package}
\title{targetesme: Collection of R functions implementing target trial emulation in the French ESME database}
\description{
Collection of R functions developped during the Alison Antoine PhD thesis (2021-2024) that helps implementing target trial emulation methodology in the French ESME database.
}
\author{
\strong{Maintainer}: Alison Antoine \email{alison.antoine@lyon.unicancer.fr}

Authors:
\itemize{
  \item Youenn Drouet \email{youenn.drouet@lyon.unicancer.fr}
}

}
\keyword{internal}
