% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetesme-package.R
\docType{package}
\name{targetesme-package}
\alias{targetesme}
\alias{targetesme-package}
\title{targetesme: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
